Changed to more reasonable build variables (#155)
* Add better acceptable code coverage range

We have 91% code coverage and the badge is still orange. Let’s be less harsh on ourselves and allow a bit more green in the codecove badge.

* Be more reasonable on patch coverage

If we are encouraging small PR’s, it’s difficult to get high patch coverage. Be more fogiving on small PR’s.

* Don’t track old verisons of golang

As we are also now tracking versions of GoCD as well as golang, there number of combinations could get quite large for travis runs, so lets’ reduce the number of golang versions we will support from 3 to 2.